Create Util Directory
