+++ android/src/main/java/com/instabug/flutter/InstabugFlutterPlugin.java
@@ -7,8 +7,6 @@ import androidx.annotation.NonNull;
 import com.instabug.flutter.generated.*;
 import com.instabug.flutter.modules.*;
 
-import java.lang.reflect.Method;
-
 import io.flutter.embedding.engine.plugins.FlutterPlugin;
 import io.flutter.plugin.common.BinaryMessenger;
 import io.flutter.plugin.common.PluginRegistry.Registrar;
@@ -40,33 +38,5 @@ public class InstabugFlutterPlugin implements FlutterPlugin {
         FeatureRequestsPigeon.FeatureRequestsHostApi.setup(messenger, new FeatureRequestsApi());
         RepliesPigeon.RepliesHostApi.setup(messenger, new RepliesApi(messenger));
         SurveysPigeon.SurveysHostApi.setup(messenger, new SurveysApi(messenger));
-    }
-
-    /**
-     * Gets the private method that matches the class, method name and parameter
-     * types given and making it accessible. For private use only.
-     * 
-     * @param clazz         the class the method is in
-     * @param methodName    the method name
-     * @param parameterType list of the parameter types of the method
-     * @return the method that matches the class, method name and param types given
-     */
-    public static Method getMethod(Class clazz, String methodName, Class... parameterType) {
-        final Method[] methods = clazz.getDeclaredMethods();
-        for (Method method : methods) {
-            if (method.getName().equals(methodName) && method.getParameterTypes().length == parameterType.length) {
-                for (int i = 0; i < parameterType.length; i++) {
-                    if (method.getParameterTypes()[i] == parameterType[i]) {
-                        if (i == method.getParameterTypes().length - 1) {
-                            method.setAccessible(true);
-                            return method;
-                        }
-                    } else {
-                        break;
-                    }
-                }
-            }
-        }
-        return null;
     }
 }
