+++ android/src/main/java/com/instabug/flutter/modules/BugReportingApi.java
@@ -7,7 +7,7 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
 import com.instabug.bug.BugReporting;
-import com.instabug.flutter.ArgsRegistry;
+import com.instabug.flutter.util.ArgsRegistry;
 import com.instabug.flutter.generated.BugReportingPigeon;
 import com.instabug.library.Feature;
 import com.instabug.library.extendedbugreport.ExtendedBugReport;
