+++ android/src/main/java/com/instabug/flutter/util/Reflection.java
@@ -1,0 +1,33 @@
+package com.instabug.flutter.util;
+
+import java.lang.reflect.Method;
+
+public class Reflection {
+    /**
+     * Gets the private method that matches the class, method name and parameter
+     * types given and making it accessible. For private use only.
+     *
+     * @param clazz         the class the method is in
+     * @param methodName    the method name
+     * @param parameterType list of the parameter types of the method
+     * @return the method that matches the class, method name and param types given
+     */
+    public static Method getMethod(Class clazz, String methodName, Class... parameterType) {
+        final Method[] methods = clazz.getDeclaredMethods();
+        for (Method method : methods) {
+            if (method.getName().equals(methodName) && method.getParameterTypes().length == parameterType.length) {
+                for (int i = 0; i < parameterType.length; i++) {
+                    if (method.getParameterTypes()[i] == parameterType[i]) {
+                        if (i == method.getParameterTypes().length - 1) {
+                            method.setAccessible(true);
+                            return method;
+                        }
+                    } else {
+                        break;
+                    }
+                }
+            }
+        }
+        return null;
+    }
+}
